+++ reqctx/context.go
@@ -32,7 +32,10 @@ var cancelFunc = contextKeyType(9)
 var sessionKey = contextKeyType(10)
 var spkgKey = contextKeyType(11)
 var activeRequestsHandlerKey = contextKeyType(12)
-var ethCallFallbackToLatestDuration = contextKeyType(13)
+var includePartialBlocksKey = contextKeyType(13)
+var partialBlocksOnlyKey = contextKeyType(14)
+var ethCallFallbackToLatestDuration = contextKeyType(15)
+var ethCallUseBlockNumberDuration = contextKeyType(16)
 
 func WithSpkg(ctx context.Context, pkg *pbsubstreams.Package) context.Context {
 	return context.WithValue(ctx, spkgKey, pkg)
@@ -56,6 +59,30 @@ func ActiveRequestsHandler(ctx context.Context) *active_requests.ActiveRequestsH
 		return t
 	}
 	return nil
+}
+
+func WithPartialBlocksOnly(ctx context.Context, partialBlocksOnly bool) context.Context {
+	return context.WithValue(ctx, partialBlocksOnlyKey, partialBlocksOnly)
+}
+
+func PartialBlocksOnly(ctx context.Context) bool {
+	val := ctx.Value(partialBlocksOnlyKey)
+	if b, ok := val.(bool); ok {
+		return b
+	}
+	return false
+}
+
+func WithIncludePartialBlocks(ctx context.Context, includePartialblocks bool) context.Context {
+	return context.WithValue(ctx, includePartialBlocksKey, includePartialblocks)
+}
+
+func IncludePartialBlocks(ctx context.Context) bool {
+	val := ctx.Value(includePartialBlocksKey)
+	if b, ok := val.(bool); ok {
+		return b
+	}
+	return false
 }
 
 func WithCancelFunc(ctx context.Context, f context.CancelCauseFunc) context.Context {