+++ tests_e2e/partial_blocks_store/src/pb/mod.rs
@@ -1,0 +1,19 @@
+// @generated
+pub mod sf {
+    pub mod acme {
+        pub mod r#type {
+            // @@protoc_insertion_point(attribute:sf.acme.type.v1)
+            pub mod v1 {
+                include!("sf.acme.type.v1.rs");
+                // @@protoc_insertion_point(sf.acme.type.v1)
+            }
+        }
+    }
+}
+pub mod test {
+    // @@protoc_insertion_point(attribute:test.output)
+    pub mod output {
+        include!("test.output.rs");
+        // @@protoc_insertion_point(test.output)
+    }
+}